TIP-860: Move the rest of normalizer of VersioningBundle
