+++ src/Akeneo/Pim/Enrichment/Component/Product/Normalizer/Versioning/Product/MetricNormalizer.php
@@ -1,8 +1,9 @@
 <?php
 
-namespace Akeneo\Tool\Bundle\VersioningBundle\Normalizer\Flat;
+namespace Akeneo\Pim\Enrichment\Component\Product\Normalizer\Versioning\Product;
 
 use Akeneo\Pim\Enrichment\Component\Product\Model\MetricInterface;
+use Akeneo\Pim\Enrichment\Component\Product\Normalizer\Versioning\Product\AbstractValueDataNormalizer;
 
 /**
  * Normalize a metric data
