+++ src/Akeneo/Pim/Enrichment/Component/spec/Product/Normalizer/Versioning/Product/DateTimeNormalizerSpec.php
@@ -1,14 +1,15 @@
 <?php
 
-namespace spec\Akeneo\Tool\Bundle\VersioningBundle\Normalizer\Flat;
+namespace spec\Akeneo\Pim\Enrichment\Component\Product\Normalizer\Versioning\Product;
 
 use PhpSpec\ObjectBehavior;
+use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
 
 class DateTimeNormalizerSpec extends ObjectBehavior
 {
     function it_is_a_normalizer()
     {
-        $this->shouldBeAnInstanceOf('Symfony\Component\Serializer\Normalizer\NormalizerInterface');
+        $this->shouldBeAnInstanceOf(NormalizerInterface::class);
     }
 
     function it_supports_csv_normalization_of_date()
