+++ src/Akeneo/Platform/Bundle/ImportExportBundle/Resources/config/normalizers.yml
@@ -1,6 +1,7 @@
 parameters:
     pim_import_export.normalizer.job_execution.class:  Akeneo\Platform\Bundle\ImportExportBundle\Normalizer\JobExecutionNormalizer
     pim_import_export.normalizer.step_execution.class: Akeneo\Platform\Bundle\ImportExportBundle\Normalizer\StepExecutionNormalizer
+    pim_serializer.normalizer.flat.job_instance.class:      Akeneo\Platform\Bundle\ImportExportBundle\Normalizer\Versioning\JobInstanceNormalizer
 
 services:
     pim_import_export.normalizer.job_execution:
@@ -21,4 +22,9 @@ services:
         tags:
             - { name: pim_serializer.normalizer, priority: 90 }
             - { name: pim_standard_format_serializer.normalizer, priority: 90 }
+            - { name: pim_versioning.serializer.normalizer, priority: 90 }
+
+    pim_versioning.serializer.normalizer.flat.job_instance:
+        class: '%pim_serializer.normalizer.flat.job_instance.class%'
+        tags:
             - { name: pim_versioning.serializer.normalizer, priority: 90 }
