+++ src/Akeneo/Tool/Bundle/VersioningBundle/DependencyInjection/AkeneoVersioningExtension.php
@@ -48,7 +48,6 @@ class AkeneoVersioningExtension extends Extension
     protected function loadSerializerConfig(array $configs, ContainerBuilder $container)
     {
         $loader = new Loader\YamlFileLoader($container, new FileLocator(__DIR__.'/../Resources/config/serializer'));
-        $loader->load('flat.yml');
         $loader->load('serializer.yml');
     }
 }
