--- src/Akeneo/Tool/Bundle/VersioningBundle/Resources/config/serializer/flat.yml
@@ -1,59 +1,0 @@
-parameters:
-    # Normalizers
-    pim_serializer.normalizer.flat.group.class:             Akeneo\Tool\Bundle\VersioningBundle\Normalizer\Flat\GroupNormalizer
-    pim_serializer.normalizer.flat.job_instance.class:      Akeneo\Tool\Bundle\VersioningBundle\Normalizer\Flat\JobInstanceNormalizer
-    pim_serializer.normalizer.flat.label_translation.class: Akeneo\Tool\Bundle\VersioningBundle\Normalizer\Flat\TranslationNormalizer
-    pim_serializer.normalizer.flat.collection.class:        Akeneo\Tool\Bundle\VersioningBundle\Normalizer\Flat\CollectionNormalizer
-    pim_serializer.normalizer.flat.datetime.class:          Akeneo\Tool\Bundle\VersioningBundle\Normalizer\Flat\DateTimeNormalizer
-    pim_serializer.normalizer.flat.metric.class:            Akeneo\Tool\Bundle\VersioningBundle\Normalizer\Flat\MetricNormalizer
-    pim_serializer.normalizer.flat.price.class:             Akeneo\Tool\Bundle\VersioningBundle\Normalizer\Flat\PriceNormalizer
-    pim_serializer.normalizer.flat.file.class:              Akeneo\Tool\Bundle\VersioningBundle\Normalizer\Flat\FileNormalizer
-
-services:
-    pim_versioning.serializer.normalizer.flat.group:
-        class: '%pim_serializer.normalizer.flat.group.class%'
-        arguments:
-            - '@pim_catalog.normalizer.standard.group'
-            - '@pim_versioning.serializer.normalizer.flat.label_translation'
-        tags:
-            - { name: pim_versioning.serializer.normalizer, priority: 90 }
-
-    pim_versioning.serializer.normalizer.flat.job_instance:
-        class: '%pim_serializer.normalizer.flat.job_instance.class%'
-        tags:
-            - { name: pim_versioning.serializer.normalizer, priority: 90 }
-
-    pim_versioning.serializer.normalizer.flat.label_translation:
-        class: '%pim_serializer.normalizer.flat.label_translation.class%'
-        tags:
-            - { name: pim_versioning.serializer.normalizer, priority: 90 }
-
-    pim_versioning.serializer.normalizer.flat.collection:
-        class: '%pim_serializer.normalizer.flat.collection.class%'
-        tags:
-            - { name: pim_versioning.serializer.normalizer, priority: 90 }
-
-    pim_versioning.serializer.normalizer.flat.metric:
-        class: '%pim_serializer.normalizer.flat.metric.class%'
-        tags:
-            - { name: pim_versioning.serializer.normalizer, priority: 90 }
-
-    pim_versioning.serializer.normalizer.flat.datetime:
-        class: '%pim_serializer.normalizer.flat.datetime.class%'
-        tags:
-            - { name: pim_versioning.serializer.normalizer, priority: 90 }
-
-    pim_versioning.serializer.normalizer.flat.price:
-        class: '%pim_serializer.normalizer.flat.price.class%'
-        tags:
-            - { name: pim_versioning.serializer.normalizer, priority: 90 }
-
-    pim_versioning.serializer.normalizer.flat.file:
-        class: '%pim_serializer.normalizer.flat.file.class%'
-        tags:
-            - { name: pim_versioning.serializer.normalizer, priority: 90 }
-
-    pim_reference_data.normalizer.flat.reference_data:
-        class: 'Akeneo\Tool\Bundle\VersioningBundle\Normalizer\Flat\ReferenceDataNormalizer'
-        tags:
-            - { name: pim_versioning.serializer.normalizer, priority: 90 }
